wip: class names refactored
