+++ src/team-mapper-api/Program.cs
@@ -17,8 +17,8 @@ builder.Services.AddScoped(typeof(IRepository<>), typeof(Repository<>));
 builder.Services.AddScoped<IWorkItemsManager, WorkItemsManager>();
 builder.Services.AddScoped<IWorkItemService, WorkItemService>();
 builder.Services.AddScoped<IPollyPolicyWrapper, PollyPolicyWrapper>();
-builder.Services.AddScoped<IExpiringWorkItemsCronService, ExpiringWorkItemsCronService>();
-builder.Services.AddHostedService<ConsumeScopedExpiringWorkItems>();
+builder.Services.AddScoped<IExpiringWorkItemsService, ExpiringWorkItemsCronService>();
+builder.Services.AddHostedService<ExpiringWorkItemsCronService>();
 
 var connectionString = builder.Configuration.GetConnectionString("TeamMapperDb");
 builder.Services.AddHealthChecks();
